requirements.txt
streamlit
python-dotenv
pandas
pymongo
plotly
XlsxWriter
